fix: resolve recovery CI integration failures
